Use factory for code node
